Lost progress trying to push to a detached branch.
Luckily the hardest work - the Javascript files were kept locally; however, I have to set up the redirects and the new routes again. That's okay - essentially all of the pull logic has been saved to the javascript files, but changes made to the toy model and the controller have been lost. :( It will take significantly less time to get that back together - the JavaScript is the most important aspect here.